requirements.txt
Flask==3.0.0
python-docx==1.1.0
Werkzeug==3.0.1
gunicorn==21.2.0
